Half of the projDetails page is now editable
I have made the description, title and picture editable. I have not
done the filters or the open/closed part yet.
This has not been tested on the database version, only on my local
version of the website. Something is wrong with the backend (at least
on my machince) that is not allowing the site to load so I did not get
a chance to make sure it works. If you see any problems please let me
know